+++ src/Sanitiser.Core/Services/ISanitisationService.cs
@@ -4,7 +4,7 @@ namespace Umbraco.Community.Sanitiser.services;
 
 public interface ISanitisationService
 {
-    public Task Sanitise(SanitisersCollection sanitisers);
+    public Task Sanitise(SanitisersCollection sanitisers, CancellationToken cancellationToken = default);
 
     public bool IsEnabled();
 }